pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vision_instruct_recipes"
version = "0.1.0"
description = "Scalable Visual Instruction Tuning Pipeline"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "tokenizers>=0.15.0",
    "accelerate>=0.25.0",
    "peft>=0.7.0",
    "bitsandbytes>=0.41.0",
    "hydra-core>=1.3.0",
    "wandb>=0.16.0",
    "pillow>=10.0.0",
    "deepspeed>=0.12.0",
    "scipy",
    "einops"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
    "isort"
]
web = [
    "gradio>=4.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.black]
line-length = 120
